Much closer to working decoder, adds PPM export
This version of the decoder is _much_ closer to working than it was
last time. A key problem was the check for the full-byte tags. During
my previous refactor from b"0xC0" -> int("C0", 16), I missed the RGB
and RGBA tags in the _decode() if/elif block. This was causing each
RGB(A) tag to be interpreted as a Run, which led to completely wrong
file sizes, and incorrect decodes.

Also added a "write_ppm()" function to QoiDecoder to allow for quick
spot checks of results. Writing to Netpbm format was even easier than
trying to come up with some bespoke serializer/deserializer.